something strange with portal
