+++ packages/commonwealth/client/scripts/views/components/component_kit/cw_growl.tsx
@@ -7,29 +7,36 @@ import 'components/component_kit/cw_growl.scss';
 import { CWCard } from './cw_card';
 import { getClasses } from './helpers';
 import { ComponentType } from './types';
+import { CWPortal } from './cw_portal';
 
 type GrowlPosition = 'bottom-left' | 'bottom-right';
 
 type GrowlAttrs = {
   onclose: () => void;
+  disabled: boolean;
   position: GrowlPosition;
   className?: string;
 };
 
 export class CWGrowl implements m.ClassComponent<GrowlAttrs> {
   view(vnode) {
-    const { className, position } = vnode.attrs;
+    const { className, position, disabled } = vnode.attrs;
+    console.log(vnode.children);
     return (
-      <div
-        class={getClasses<{ className?: string; position: GrowlPosition }>(
-          { className, position },
-          ComponentType.Growl
-        )}
-      >
-        <CWCard className="growl-card" elevation="elevation-3" interactive>
-          {vnode.children}
-        </CWCard>
-      </div>
+      !disabled && (
+        <CWPortal>
+          <div
+            class={getClasses<{ className?: string; position: GrowlPosition }>(
+              { className, position },
+              ComponentType.Growl
+            )}
+          >
+            <CWCard className="growl-card" elevation="elevation-3" interactive>
+              {vnode.children}
+            </CWCard>
+          </div>
+        </CWPortal>
+      )
     );
   }
 }
